Change oper_name_t to 16-bit integer. Added support for 16- and 32-bit integers in  read_positive_integer/read_positive_integer_array functions.

+++ cpp/lib_moment/integer_types.h
@@ -23,7 +23,7 @@ namespace Moment {
     /**
      * For enumerating operators.
      */
-    using oper_name_t = int64_t;
+    using oper_name_t = int16_t;
 
     /**
      * For enumerating symbolic algebra elements
